chore: Moves file input to internal
